like and favorites func added in controller

+++ src/main/java/com/coding/meal_kit/controllers/MealController.java
@@ -1,6 +1,7 @@
 package com.coding.meal_kit.controllers;
 
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import com.coding.meal_kit.models.Areas;
 import com.coding.meal_kit.models.CountryMeals;
+import com.coding.meal_kit.models.Meal;
 import com.coding.meal_kit.models.Meals;
 import com.coding.meal_kit.models.User;
 import com.coding.meal_kit.services.MealService;
@@ -45,33 +47,41 @@ public class MealController {
 
 	@GetMapping("/")
 	public String home(Model model) {
-		Meals meals = mealService.getRandomMeal();
-		// System.out.println(areas.getAreas().toString());
-		model.addAttribute("apiData", meals);
+		// userId
 		model.addAttribute("apiAreas", this.areas);
 		return "/meal/home.jsp";
 	}
 
 	@GetMapping("/getMeal/{id}")
-	public String showMealByCountry(@PathVariable("id") String id, Model model) {
-		Meals meals = mealService.getMealbyID(id);
-		model.addAttribute("apiData", meals);
-		model.addAttribute("apiAreas", this.areas);
-		return "/meal/home.jsp";
-	}
-
-
-	@GetMapping("/details")
-	public String details(Model model) {
-		Meals meals = mealService.getRandomMeal();
-		// System.out.println(areas.getAreas().toString());
-		model.addAttribute("apiData", meals);
+	public String showMealByCountry(@PathVariable("id") Long id, Model model, HttpSession session) {
+		if (session.getAttribute("userId") == null)
+			return "redirect:login/";
+		Meal meal = mealService.getMealbyID(id);
+		User user = this.uService.findById((Long) session.getAttribute("userId"));
+		model.addAttribute("meal", meal);
+		model.addAttribute("user",user);
 		model.addAttribute("apiAreas", this.areas);
 		return "/meal/details.jsp";
 	}
 
+
+	@GetMapping("/details")
+	public String details(Model model, HttpSession session) {
+		if (session.getAttribute("userId") == null)
+			return "redirect:/login/";
+		Meal meal = mealService.getRandomMeal();
+		User user = this.uService.findById((Long) session.getAttribute("userId"));
+		model.addAttribute("user",user);
+		model.addAttribute("meal", meal);
+		model.addAttribute("apiAreas", this.areas);
+		
+		return "/meal/details.jsp";
+	}
+
 	@GetMapping("/searchByCountry")
-	public String getMealByCountry(@RequestParam("country") String country, Model model) {
+	public String getMealByCountry(@RequestParam("country") String country, Model model, HttpSession session) {
+		if (session.getAttribute("userId") == null)
+			return "redirect:login/";
 		CountryMeals meals = mealService.getMealbyCountry(country);
 		model.addAttribute("apiAreas", this.areas);
 		model.addAttribute("apiData", meals);
@@ -79,7 +89,9 @@ public class MealController {
 	}
 	
 	@GetMapping("/search")
-	public String search(@RequestParam("term") String term, @RequestParam("by") String by, Model model) {
+	public String search(@RequestParam("term") String term, @RequestParam("by") String by, Model model, HttpSession session) {
+		if (session.getAttribute("userId") == null)
+			return "redirect:login/";
 		
 		if (by.equals("name")) {
 			Meals meal = mealService.getMealbyName(term);
@@ -106,14 +118,18 @@ public class MealController {
 	}
 	
 	@GetMapping("/profile/{id}")
-	public String profile(@PathVariable("id")Long id, Model model) {
+	public String profile(@PathVariable("id")Long id, Model model, HttpSession session) {
+		if (session.getAttribute("userId") == null)
+			return "redirect:login/";
 		model.addAttribute("pUser", uService.findById(id));
 		return "/meal/profile.jsp";
 		
 	}
 	
 	@GetMapping("profile/{id}/edit")
-	public String edit( @ModelAttribute("Edit") User user, @PathVariable("id") Long id, Model model) {
+	public String edit( @ModelAttribute("Edit") User user, @PathVariable("id") Long id, Model model, HttpSession session) {
+		if (session.getAttribute("userId") == null)
+			return "redirect:login/";
 
 		model.addAttribute("EditOne", uService.findById(id));
 
@@ -124,7 +140,9 @@ public class MealController {
 
 	@PutMapping("/update/{id}")
 	public String update(@Valid @ModelAttribute("Edit") User user, BindingResult results, @PathVariable("id") Long id,
-			Model model) {
+			Model model, HttpSession session) {
+		if (session.getAttribute("userId") == null)
+			return "redirect:login/";
 		if (results.hasErrors()) {
 			model.addAttribute("EditOne", uService.findById(id));
 			return "/meal/edit.jsp";
@@ -137,10 +155,62 @@ public class MealController {
 	}
 
 	@DeleteMapping("/delete/{id}")
-	public String deleteRoom(@PathVariable("id") Long id) {
+	public String deleteRoom(@PathVariable("id") Long id, HttpSession session) {
+		if (session.getAttribute("userId") == null)
+			return "redirect:login/";
 		uService.deleteU(id);
 		return "redirect:/";
 	}
+	
+	// Like
+		@GetMapping("/{id}/like")
+		public String like(HttpSession session, @PathVariable("id") Long id) {
+			
+			// Check if there is any active user session.
+			if (session.getAttribute("userId") == null)
+				return "redirect:login/";
+			User user = this.uService.findById((Long) session.getAttribute("userId"));
+			Meal meal = this.mealService.getMealbyID(id);// idMeal
+			this.mealService.likeMeal(user, meal);
+			return "redirect:/getMeal/"+id;
+		}
+
+		// Unlike
+		@GetMapping("/{id}/unlike")
+		public String unlike(HttpSession session, @PathVariable("id") Long id) {
+			// Check if there is any active user session.
+			if (session.getAttribute("userId") == null)
+				return "redirect:login/";
+			User user = this.uService.findById((Long) session.getAttribute("userId"));
+			Meal meal = this.mealService.getMealbyID(id);// idMeal
+			this.mealService.unlikeMeal(user, meal);
+			return "redirect:/getMeal/"+id;
+		}
+		
+		// Add Favorite
+				@GetMapping("/{id}/addFavorites")
+				public String addFavorite(HttpSession session, @PathVariable("id") Long id) {
+					
+					// Check if there is any active user session.
+					if (session.getAttribute("userId") == null)
+						return "redirect:login/";
+					User user = this.uService.findById((Long) session.getAttribute("userId"));
+					Meal meal = this.mealService.getMealbyID(id);// idMeal
+					this.mealService.addFavoriteMeal(user, meal);
+					return "redirect:/getMeal/"+id;
+				}
+		
+				// Remove From Favorites
+				@GetMapping("/{id}/removeFavorite")
+				public String removeFavorite(HttpSession session, @PathVariable("id") Long id) {
+					// Check if there is any active user session.
+					if (session.getAttribute("userId") == null)
+						return "redirect:login/";
+					User user = this.uService.findById((Long) session.getAttribute("userId"));
+					Meal meal = this.mealService.getMealbyID(id);// idMeal
+					this.mealService.removeFromFavorites(user, meal);
+					return "redirect:/getMeal/"+id;
+				}
 
 
 }
